+++ CHANGELOG.md
@@ -1,2 +1,10 @@
 
-* initial release
+Seperate rolls for coins and bijous added:
+* the old commands (/bijouroll [need|greed|pass]) still work the same as before
+* by default the roll behaviour for coins is locked to the behaviour for bijous
+* to unlock the coin rolls, use /zgroll [c|coin] [need|greed|pass]
+* to lock the coins again, use /zgroll [c|coin] lock
+
+Command line parameters (like need, greed and pass) are no longer case sensitive.
+
+Versioning in ZGBijouRoller.toc fixed.